(registration-api) Refactor ApiDocumentBuilder to Spring Service, to use dependency injection instead passing through services
